+++ app/src/main/java/edu/upc/eetac/dsa/calculator/Operation.java
@@ -1,5 +1,5 @@
 package edu.upc.eetac.dsa.calculator;
 
 public interface Operation {
-    public int operation(int a, int b);
+    public double operation(double a, double b);
 }
