+++ app/src/main/java/edu/upc/eetac/dsa/calculator/Product.java
@@ -2,7 +2,7 @@ package edu.upc.eetac.dsa.calculator;
 
 public class Product implements Operation{
     @Override
-    public int operation(int a, int b) {
+    public double operation(double a, double b) {
         return a*b;
     }
 }
